Handle mapping of steps with custom hook instead of with redux object
