+++ src/types/index.ts
@@ -26,6 +26,19 @@ export type User = {
 
 export type Schema = "study" | "dac" | "policy" | "sample" | "experiment" | "run" | "analysis" | "dataset"
 
+export type StepItemObject = { id: string; displayTitle: string; objectData?: ObjectInsideFolderWithTags }
+
+export type StepObject = {
+  label: string
+  stepItems?: {
+    objectType: string
+    label: string
+    objects?: { ready?: StepItemObject[]; drafts?: StepItemObject[] }
+  }[]
+  actionButtonText?: string
+  disabled?: boolean
+}
+
 export type ObjectDetails = {
   accessionId: string
   lastModified: string
@@ -83,6 +96,8 @@ export type FolderDataFromForm = {
   name: string
   description: string
 }
+
+export type SubmissionFolder = FolderDetailsWithId & { doiInfo: DoiFormDetails }
 
 export type CreateFolderFormRef = { current: HTMLElement | null } | undefined
